+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Made PHP8.1 minimum version
+
 ## v0.2.8 (2023-10-18)
 * Updated Genesis dependency