+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.2.9 (2024-03-15)
+* Fixed CSS as entry point
+* Made PHP8.1 minimum version
+
 ## v0.2.8 (2023-10-18)
 * Updated Genesis dependency